address notes
