--- app/src/redux/protocol-analysis/constants.ts
@@ -1,5 +1,0 @@
-export const OPEN_PYTHON_DIRECTORY: 'protocol-analysis:OPEN_PYTHON_DIRECTORY' =
-  'protocol-analysis:OPEN_PYTHON_DIRECTORY'
-
-export const CHANGE_PYTHON_PATH_OVERRIDE: 'protocol-analysis:CHANGE_PYTHON_PATH_OVERRIDE' =
-  'protocol-analysis:CHANGE_PYTHON_PATH_OVERRIDE'
