+++ app/src/redux/protocol-analysis/index.ts
@@ -1,3 +1,1 @@
 export * from './actions'
-export * from './constants'
-export * from './types'
